Refactored project and changes tests to rustdoc comments

+++ src/linear/vector.rs
@@ -18,8 +18,8 @@ use std::ops::{Add, Div, Mul, Range, Sub};
 ///
 #[derive( Copy, Clone, Debug )]
 pub struct Vec2 {
-	pub x: f64,
-	pub y: f64,
+	pub x: f32,
+	pub y: f32,
 	coord_system: CoordSystem,
 }
 
@@ -161,7 +161,7 @@ impl Vec2 {
 	/// assert!( ( ( a.x - 5.6 ) < 0.0000000001 ) && ( ( a.y - 7.2 ) < 0.0000000001 ) );
 	/// ```
 	///
-	pub fn create( x: &f64, y: &f64,   coord_system: &CoordSystem ) -> Vec2 {
+	pub fn create( x: &f32, y: &f32,   coord_system: &CoordSystem ) -> Vec2 {
 		let mut temp = Vec2 {
 			x: *x,
 			y: *y,
@@ -181,7 +181,7 @@ impl Vec2 {
 	/// 
 	/// ```
 	///
-	pub fn create_random( range_x: &Range<f64>, range_y: &Range<f64>,   coord_system: &CoordSystem ) -> Vec2 {
+	pub fn create_random( range_x: &Range<f32>, range_y: &Range<f32>,   coord_system: &CoordSystem ) -> Vec2 {
 		let mut temp = Vec2 {
 			x: rand::thread_rng().gen_range( range_x.clone() ),
 			y: rand::thread_rng().gen_range( range_y.clone() ),
@@ -201,7 +201,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn from_angle( theta: &f64 ) -> Vec2 {
+	pub fn from_angle( theta: &f32 ) -> Vec2 {
 		let mut temp = Vec2::create( &1.0, theta, &POLAR );
 		temp.swap_system( CARTESIAN );
 		temp
@@ -212,7 +212,7 @@ impl Vec2 {
 	/// # Examples
 	///
 	/// ```
-	/// use std::f64::consts::TAU;
+	/// use std::f32::consts::TAU;
 	/// use miscmath::linear::{ CoordSystem::*, vector::Vec2 };
 	///
 	/// let mut a = Vec2::random_unit( &( 0.0..TAU ) );
@@ -221,7 +221,7 @@ impl Vec2 {
 	/// assert!( ( ( a.x - 1.0 ) < 0.0000000001 ) && ( a.y < TAU ) );
 	/// ```
 	///
-	pub fn random_unit( range: &Range<f64> ) -> Vec2 {
+	pub fn random_unit( range: &Range<f32> ) -> Vec2 {
 		let mut temp = Vec2::create( &1.0, &rand::thread_rng().gen_range( range.clone() ), &POLAR );
 		temp.swap_system( CARTESIAN );
 		temp
@@ -277,7 +277,7 @@ impl Vec2 {
 	/// assert!( ( angle - 1.4731481877 ) < 0.0000000001 );
 	/// ```
 	///
-	pub fn angle_between( &self, rhs: &Vec2 ) -> f64 {
+	pub fn angle_between( &self, rhs: &Vec2 ) -> f32 {
 		let dot = self.dot( &rhs );
 		( dot / ( self.mag() * rhs.mag() ) ).acos()
 	}
@@ -309,7 +309,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn dist( &self, rhs: &Vec2 ) -> f64 {
+	pub fn dist( &self, rhs: &Vec2 ) -> f32 {
 		( ( self.x - rhs.x ).powi(2) ) + ( ( self.y - rhs.y ).powi(2) )
 	}
 	
@@ -337,7 +337,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn dot( &self, rhs: &Vec2 ) -> f64 {
+	pub fn dot( &self, rhs: &Vec2 ) -> f32 {
 		( self.x * rhs.x ) + ( self.y * rhs.y )
 	}
 	
@@ -351,7 +351,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn div( &mut self, rhs: &f64 ) {
+	pub fn div( &mut self, rhs: &f32 ) {
 		self.x /= rhs;
 		self.y /= rhs;
 	}
@@ -387,7 +387,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn mag( &self ) -> f64 {
+	pub fn mag( &self ) -> f32 {
 		( ( self.x ).powi(2) ).sqrt() + ( self.y ).powi(2)
 	}
 	
@@ -401,7 +401,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn mag_sq( &self ) -> f64 {
+	pub fn mag_sq( &self ) -> f32 {
 		( self.x ).powi(2) + ( self.y ).powi(2)
 	}
 	
@@ -415,7 +415,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn mult( &mut self, rhs: &f64 ) {
+	pub fn mult( &mut self, rhs: &f32 ) {
 		self.swap_system( CARTESIAN );
 		self.x *= rhs;
 		self.y *= rhs;
@@ -447,7 +447,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn rem( &mut self, rhs: &f64 ) {
+	pub fn rem( &mut self, rhs: &f32 ) {
 		self.swap_system( CARTESIAN );
 		self.x %= rhs;
 		self.y %= rhs;
@@ -463,7 +463,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn rotate( &mut self, theta: f64 ) {
+	pub fn rotate( &mut self, theta: f32 ) {
 		self.swap_system( POLAR );
 		self.y += theta;
 		self.swap_system( CARTESIAN );
@@ -479,7 +479,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn set( &mut self, input1: &f64, input2: &f64,   coord_system: &CoordSystem ) {
+	pub fn set( &mut self, input1: &f32, input2: &f32,   coord_system: &CoordSystem ) {
 		if *coord_system == CARTESIAN {
 			self.swap_system( CARTESIAN );
 			self.x = *input1;
@@ -502,7 +502,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn set_mag( &mut self, input: &f64 ) {
+	pub fn set_mag( &mut self, input: &f32 ) {
 		self.swap_system( POLAR );
 		self.x = *input;
 		self.swap_system( CARTESIAN );
@@ -518,7 +518,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn set_theta( &mut self, input: &f64 ) {
+	pub fn set_theta( &mut self, input: &f32 ) {
 		self.swap_system( POLAR );
 		self.y = *input;
 		self.swap_system( CARTESIAN );
@@ -591,7 +591,7 @@ impl Vec2 {
 	///   
 	/// ```
 	///
-	pub fn theta( &self ) -> f64 {
+	pub fn theta( &self ) -> f32 {
 		( self.y ).atan2( self.x )
 	}
 }
